# lib/api-zod/src/generated/types/ngoVerifyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VolunteerConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NgoVerifyResponse {
  verified: boolean;
  message: string;
}